requirements.txt
# API Framework - Versioni compatibili senza Rust
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.10.12

# Basic utilities
requests==2.28.0
python-dotenv==0.19.0

# PDF processing
PyMuPDF==1.23.0
aiofiles==24.1.0

# Production server
gunicorn==20.1.0

# AI Libraries for OCR
moondream==0.1.1
transformers==4.46.3
torch>=2.0.0
torchvision>=0.15.0
einops==0.8.1
pillow>=10.0.0